refactor: move react-color code to subdirectory
