+++ src/components/color-picker/react-color/ColorPicker.tsx
