+++ src/components/color-picker/react-color/LICENSE
