+++ src/components/color-picker/react-color/common/Alpha.jsx
