+++ src/components/color-picker/react-color/common/CheckBoard.jsx
