+++ src/components/color-picker/react-color/common/EditableInput.jsx
