+++ src/components/color-picker/react-color/common/Swatch.jsx
