+++ src/components/color-picker/react-color/common/index.js
