+++ src/components/color-picker/react-color/common/interaction.jsx
