+++ src/components/color-picker/react-color/helpers/alpha.js
