+++ src/components/color-picker/react-color/helpers/checkBoard.js
