+++ src/components/color-picker/react-color/helpers/color.js
