+++ src/components/color-picker/react-color/helpers/hue.js
