+++ src/components/color-picker/react-color/helpers/saturation.js
