+++ src/components/color-picker/react-color/sketch/SketchPresetColors.tsx
